back-end/Crowd-Funding-main/requirements.txt
asgiref==3.7.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
Pillow==9.5.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.0
